+++ pyFAI/gui/utils/unitutils.py
@@ -1,6 +1,5 @@
 # coding: utf-8
 # /*##########################################################################
-#
 # Copyright (C) 2016-2018 European Synchrotron Radiation Facility
 #
 # Permission is hereby granted, free of charge, to any person obtaining a copy
@@ -21,80 +20,17 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 # THE SOFTWARE.
 #
-# ###########################################################################*/
-
-from __future__ import absolute_import
+# ############################################################################*/
 
 __authors__ = ["V. Valls"]
 __license__ = "MIT"
-__date__ = "26/11/2018"
+__date__ = "03/01/2019"
 
 
 import numpy
 import collections
 
-from silx.gui import qt
-from silx.gui.widgets.WaitingPushButton import WaitingPushButton
-from silx.gui.colors import Colormap
-
 from pyFAI import units
-
-
-def getFreeColorRange(colormap):
-    name = colormap['name']
-    colormap = Colormap(name)
-
-    # extract all hues from colormap
-    colors = colormap.getNColors()
-    hues = []
-    for c in colors:
-        c = qt.QColor.fromRgb(c[0], c[1], c[2])
-        hues.append(c.hueF())
-
-    # search the bigger empty hue range
-    current = (0, 0.0, 0.2)
-    hues = filter(lambda x: x >= 0, set(hues))
-    hues = list(sorted(hues))
-    if len(hues) > 1:
-        for i in range(len(hues)):
-            h1 = hues[i]
-            h2 = hues[(i + 1) % len(hues)]
-            if h2 < h1:
-                h2 = h2 + 1.0
-
-            diff = h2 - h1
-            if diff > 0.5:
-                diff = 1.0 - diff
-
-            if diff > current[0]:
-                current = diff, h1, h2
-    elif len(hues) == 1:
-        h = (hues[0] + 0.5) % 1.0
-        current = (0, h - 0.1, h + 0.1)
-    else:
-        pass
-
-    h1, h2 = current[1:]
-    delta = (h2 - h1) / 6.0
-
-    # move the range from the colormap
-    h1, h2 = h1 + delta, h2 - delta
-    hmin = (h1 + h2) / 2.0
-
-    # generate colors with 3 hsv control points
-    # (h1, 1, 1), (hmid, 1, 0.5), (h2, 1, 1)
-    colors = []
-    for i in range(5):
-        h = h1 + (hmin - h1) * (i / 5.0)
-        v = 0.5 + 0.1 * (5 - i)
-        c = qt.QColor.fromHsvF(h % 1.0, 1.0, v)
-        colors.append(c)
-    for i in range(5):
-        h = hmin + (h2 - hmin) * (i / 5.0)
-        v = 0.5 + 0.1 * (i)
-        c = qt.QColor.fromHsvF(h % 1.0, 1.0, v)
-        colors.append(c)
-    return colors
 
 
 def tthToRad(twoTheta, unit, wavelength=None, directDist=None):
@@ -164,42 +100,3 @@ def from2ThRad(twoTheta, unit, wavelength=None, directDist=None, ai=None):
         return beamCentre * numpy.tan(twoTheta) * 0.001
     else:
         raise ValueError("Converting from 2th to unit %s is not supported", unit)
-
-
-def createProcessingWidgetOverlay(parent):
-    """Create a widget overlay to show that the application is processing data
-    to update the plot.
-
-    :param qt.QWidget widget: Widget containing the overlay
-    :rtype: qt.QWidget
-    """
-    if hasattr(parent, "centralWidget"):
-        parent = parent.centralWidget()
-    button = WaitingPushButton(parent)
-    button.setWaiting(True)
-    button.setText("Processing...")
-    button.setDown(True)
-    position = parent.size()
-    size = button.sizeHint()
-    position = (position - size) / 2
-    rect = qt.QRect(qt.QPoint(position.width(), position.height()), size)
-    button.setGeometry(rect)
-    button.setVisible(True)
-    return button
-
-
-_trueStrings = set(["yes", "true", "1"])
-_falseStrings = set(["no", "false", "0"])
-
-
-def stringToBool(string):
-    """Returns a boolean from a string.
-
-    :raise ValueError: If the string do not contains a boolean information.
-    """
-    lower = string.lower()
-    if lower in _trueStrings:
-        return True
-    if lower in _falseStrings:
-        return False
-    raise ValueError("'%s' is not a valid boolean" % string)
